// src/app/Interfaces/Cliente.ts
export class Cliente{
    //Cliente_ID;
    Nombre_Usuario: string;
    Contraseña: string;
    Nombre: string;
    Apellidos: string;
    Correo_Electronico: string;
    Edad: number;
    Estatura: number;
    Peso: number;
    //Genero_ID: number;
    //Actividad_Fisica_ID: number;
    //Dieta_ID: number;
    //Objetivo_ID: number;
    //IMC: number;
    GEB: number;
    //ETA: number;
    //Peso_Maximo: number;
    //Peso_Minimo: number;
    //AF: number;
    //Gasto_Energetico_Total: number;
    //Tipo_Cliente_ID: number;
    //Activo: boolean;
    //Orden: number;
    //Fecha_Creacion: Date;
    //Fecha_Actualizacion: Date;
    //Usuario_ID: number;
    //Visible: boolean;
    //De_Sistema: boolean;
}